fetches user credentials from database [NodeJS]
COMMENTS:
Authenticate users by checking the user credentials against the
info stored in the database.

Responds with status 500 if the user does not exist, and also if
the supplied password does not match.

Creates the JSONWebToken with the document ID for the user in
the mongo database. (Note that mongo database generates the ID
automatically upon document creation.)

Uses async/await to authenticate users asynchronously.

Authentication has been tested via curl commands and the front-end
side of the application.